+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [9.13.0] - 2024-06-01
 ### Added
 - **feat**: `url_vid` rule as a way to validate whether or not a value is a valid video url
 - **feat**: `url_aud` rule as a way to validate whether or not a value is a valid audio url